# backend/auth-service/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "OAuth 2.0 authorization and user authentication service for the library management system.\nThis service uses opaque bearer tokens, Authorization Code with PKCE, and HttpOnly session cookies.",
        "title": "Library Management Auth API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/oauth/authorize": {
            "get": {
                "description": "Validates the client, redirect URI, state, and S256 PKCE challenge. Redirects unauthenticated users to login.",
                "tags": [
                    "OAuth"
                ],
                "summary": "Authorize OAuth client",
                "parameters": [
                    {
                        "enum": [
                            "code"
                        ],
                        "type": "string",
                        "description": "OAuth response type",
                        "name": "response_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "OAuth client ID",
                        "name": "client_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Exact registered callback URL",
                        "name": "redirect_uri",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Space-delimited scopes",
                        "name": "scope",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CSRF correlation value",
                        "name": "state",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Base64url-encoded SHA-256 PKCE challenge",
                        "name": "code_challenge",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "S256"
                        ],
                        "type": "string",
                        "description": "PKCE method",
                        "name": "code_challenge_method",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to login or callback with authorization code",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid authorization request",
                        "schema": {
                            "$ref": "#/definitions/oauth.OAuthErrorResponse"
                        }
                    }
                }
            }
        },
        "/oauth/introspect": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "RFC 7662 endpoint restricted to provisioned resource-server clients using HTTP Basic.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Introspect access token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Opaque access token",
                        "name": "token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "enum": [
                            "access_token"
                        ],
                        "type": "string",
                        "description": "Token type hint",
                        "name": "token_type_hint",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token state; inactive tokens include only active=false",
                        "schema": {
                            "$ref": "#/definitions/oauth.OAuthIntrospectionResponse"
                        }
                    },
                    "400": {
                        "description": "Token missing or request invalid",
                        "schema": {
                            "$ref": "#/definitions/oauth.OAuthErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Client authentication or authorization failed",
                        "schema": {
                            "$ref": "#/definitions/oauth.OAuthErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Token state could not be determined",
                        "schema": {
                            "$ref": "#/definitions/oauth.OAuthErrorResponse"
                        }
                    }
                }
            }
        },
        "/oauth/token": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Supports authorization_code, refresh_token, and provisioned client_credentials clients. Client authentication uses HTTP Basic.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OAuth"
                ],
                "summary": "Issue OAuth token",
                "parameters": [
                    {
                        "enum": [
                            "authorization_code",
                            "refresh_token",
                            "client_credentials"
                        ],
                        "type": "string",
                        "description": "OAuth grant type",
                        "name": "grant_type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization code; required for authorization_code",
                        "name": "code",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Exact callback URL; required for authorization_code",
                        "name": "redirect_uri",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "PKCE verifier; required for authorization_code",
                        "name": "code_verifier",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Refresh token; required for refresh_token",
                        "name": "refresh_token",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Space-delimited requested scopes",
                        "name": "scope",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token issued",
                        "schema": {
                            "$ref": "#/definitions/oauth.OAuthTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or grant",
                        "schema": {
                            "$ref": "#/definitions/oauth.OAuthErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Client authentication failed",
                        "schema": {
                            "$ref": "#/definitions/oauth.OAuthErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/oauth.APIErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "oauth.APIErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "LMS-429001"
                },
                "message": {
                    "type": "string",
                    "example": "too many authentication attempts"
                }
            }
        },
        "oauth.OAuthErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid_request"
                },
                "error_description": {
                    "type": "string",
                    "example": "the request is invalid"
                }
            }
        },
        "oauth.OAuthIntrospectionResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "client_id": {
                    "type": "string",
                    "example": "member-nextjs-web"
                },
                "exp": {
                    "type": "integer",
                    "example": 1784448900
                },
                "iat": {
                    "type": "integer",
                    "example": 1784448000
                },
                "iss": {
                    "type": "string",
                    "example": "http://localhost:8081"
                },
                "scope": {
                    "type": "string",
                    "example": "library:read"
                },
                "sub": {
                    "type": "string",
                    "example": "f81d4fae-7dec-11d0-a765-00a0c91e6bf6"
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                }
            }
        },
        "oauth.OAuthTokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "opaque-access-token"
                },
                "expires_in": {
                    "type": "integer",
                    "example": 900
                },
                "refresh_token": {
                    "type": "string",
                    "example": "opaque-refresh-token"
                },
                "scope": {
                    "type": "string",
                    "example": "library:read library:write"
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Prefix access tokens with \"Bearer \".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}